backend/requirements.txt
fastapi[all]
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
